requirements.txt
python-telegram-bot
openai
feedparser
apscheduler
python-dotenv
telethon 